pkg/ast: fix gep to handle Geode struct types
Since ir.NewGetElementPtr is unaware of Geode types,
introduce an indirection through the ast.gep function
which only present the underlying LLVM IR element type
to ir.NewGetElementPtr.

Note, this is note the prettiest approach, but seems
to work. I'd be glad to discuss alternatives that may
be cleaner than this.