Update to v0.3.0 of llir/llvm.
There were two main changes of llir/llvm that required updates in Geode,
namely:

1) The sanity type-checks proposed in llir/llvm#65 and added in PR
   llir/llvm#72.
2) The API update of NewLoad, NewStore, NewGetElementPtr which added
   a elemType argument in preparation for opaque pointer types of the
   official LLVM project (see llir/llvm#112).

To resolve 1), an Equal method was added to StructType and SliceType of
gtypes.

To resolve 2) the element type was passed as the first argument to
NewLoad, NewStore and NewGetElementPtr.

Note, we've definitely noticed that working with user-defined types is
difficult and does not work well with the extensive type-switching used
in llir/llvm. The user-defined types (structs with names and slices) of
Geode have been very helpful to get a better understanding of this
limitation of llir/llvm. This has been outlined in https://github.com/llir/llvm/issues/59#issuecomment-476271677
and we hope to get a better design for sumtypes to allow user-defined
types, values, constants, instructions, etc.

One step at the time :)

I'm still very happy that you started the Geode project Nick as I've
learnt a lot from it and continue to learn more.

Wish you all the best and a most lovely start of the new year.

Cheers,
Robin